requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=4.9.3
pandas>=2.0.3
numpy>=1.24.3
yfinance>=0.2.28
akshare>=1.12.0
TA-Lib>=0.4.28
python-dotenv>=1.0.0
jinja2>=3.1.2
schedule>=1.2.0
openai>=1.3.0
python-dateutil>=2.8.2
pytz>=2023.3
scipy>=1.11.0
email-validator>=2.1.0
